backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["goncalofonseca <goncalofnsc@gmail.com>"]
#readme = "README.md"
#package-mode = false  # Use this for simple scripts

[tool.poetry.dependencies]
python = "^3.11"
boto3 = "^1.38.26"
notion-client = "^2.3.0"
notion2md = "^2.9.0"
pinecone = "^7.0.2"
ulid-py = "^1.1.0"
anthropic = "^0.29.0"
google-api-python-client = "^2.174.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.2"
opentelemetry-api = "^1.34.1"
opentelemetry-sdk = "^1.34.1"
opentelemetry-instrumentation-botocore = "^0.55b1"

[tool.poetry.scripts]
run-app = "main:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry-plugin-lambda-build]
layer_artifact_path = "artifacts/layer.zip"
layer_install_dir = "python"
function_artifact_path = "artifacts/function.zip"

